feat(node): Add test accounts
