+++ Code/node-src/jest.config.js
@@ -1,10 +1,11 @@
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  testMatch: [ "**/__tests__/**/*.ts?(x)", "**/?(*.)+(spec|test).ts?(x)" ],
-  verbose: true,
-  collectCoverage: true,
-  coverageDirectory: 'coverage',
-  setupFilesAfterEnv: ['./tests/setup.ts'],
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    testMatch: ['**/__tests__/**/*.ts?(x)', '**/?(*.)+(spec|test).ts?(x)'],
+    modulePathIgnorePatterns: ['./app/index.ts'],
+    verbose: true,
+    collectCoverage: true,
+    coverageDirectory: 'coverage',
+    setupFilesAfterEnv: ['./tests/setup.ts'],
 };
