+++ Code/node-src/tests/classgroup.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import { app } from '../app';
+import { app } from '../app/app';
 
 /**
  * Test API routes and expected responses.
